♻️(DocNavItem.svelte): recursive component
